Close milestone 6 group session flow and align roadmap
