+++ app/[locale]/sessions/[sessionId]/page.tsx
@@ -107,7 +107,7 @@ export default async function SessionPage({ params, searchParams }: SessionPageP
               {t('startSession')}
             </SubmitButton>
           </form>
-          <Link href="/dashboard?view=sessions" className="button-ghost mt-4 px-4 py-2 text-sm text-slate-500">
+          <Link href={`/groups/${data.group.id}`} className="button-ghost mt-4 px-4 py-2 text-sm text-slate-500">
             {t('quitSession')}
           </Link>
         </section>
@@ -154,10 +154,10 @@ export default async function SessionPage({ params, searchParams }: SessionPageP
         <FeedbackBanner message={searchParams.feedbackMessage} tone={searchParams.feedbackTone} />
         <header className="sticky top-0 z-20 border-b border-white/[0.07] bg-background/95 backdrop-blur">
           <div className="mx-auto flex h-16 w-full max-w-[700px] items-center justify-between px-4">
-            <Link href="/dashboard?view=sessions" className="text-sm font-bold text-slate-500 hover:text-white">
+            <Link href={`/groups/${data.group.id}`} className="text-sm font-bold text-slate-500 hover:text-white">
               <span className="inline-flex items-center gap-2">
                 <ArrowLeft className="h-4 w-4" aria-hidden="true" />
-                Dashboard
+                {data.group.name}
               </span>
             </Link>
             <p className="text-lg font-extrabold text-white">{data.session.name ?? data.group.name} — {t('reviewShort')}</p>
@@ -261,7 +261,7 @@ export default async function SessionPage({ params, searchParams }: SessionPageP
       <FeedbackBanner message={searchParams.feedbackMessage} tone={searchParams.feedbackTone} />
       <header className="border-b border-white/[0.07]">
         <div className="mx-auto flex h-16 w-full max-w-[560px] items-center justify-between px-4">
-          <Link href="/dashboard?view=sessions" className="text-slate-500 hover:text-white">
+          <Link href={`/groups/${data.group.id}`} className="text-slate-500 hover:text-white">
             <ArrowLeft className="h-4 w-4" aria-hidden="true" />
           </Link>
           <div className="text-center">
